Add unit tests package and migrate test_color.kt to unit testing
